Cleaned up the functions in the .js files
cleaned up the move functions in the
map file... cleaned up the formatting in the top
half of the map file... cleaned up the functions
and formatting in the hero file... created a
separate file for the map cells... added a testing
line of code to the energy.js file to see if we
can use the function without explicitly including
that file.

Still need to implement some member functions in
the map and hero class to get the move functions
to work properly.  Still need to test the new
funtions.